Split components for better component management
