+++ src/components/UI/ErrorScreen.tsx
@@ -1,0 +1,45 @@
+import React from 'react';
+import { motion } from 'framer-motion';
+import { AlertCircle } from 'lucide-react';
+
+interface ErrorScreenProps {
+  error: string;
+  onRetry: () => void;
+  title?: string;
+}
+
+const ErrorScreen: React.FC<ErrorScreenProps> = ({ 
+  error, 
+  onRetry, 
+  title = "Unable to Load Hotels" 
+}) => {
+  return (
+    <div className="h-screen flex items-center justify-center bg-gradient-to-br from-red-50 to-pink-100">
+      <motion.div
+        initial={{ opacity: 0, y: 20 }}
+        animate={{ opacity: 1, y: 0 }}
+        className="text-center space-y-6 max-w-md mx-auto p-6"
+      >
+        <div className="w-16 h-16 mx-auto bg-red-100 rounded-full flex items-center justify-center">
+          <AlertCircle className="w-8 h-8 text-red-600" />
+        </div>
+        
+        <div className="space-y-2">
+          <h2 className="text-2xl font-bold text-gray-900">{title}</h2>
+          <p className="text-gray-600">{error}</p>
+        </div>
+        
+        <motion.button
+          onClick={onRetry}
+          className="px-6 py-3 bg-blue-600 text-white rounded-lg hover:bg-blue-700 transition-colors"
+          whileHover={{ scale: 1.05 }}
+          whileTap={{ scale: 0.95 }}
+        >
+          Try Again
+        </motion.button>
+      </motion.div>
+    </div>
+  );
+};
+
+export default ErrorScreen;
