+++ src/main.tsx
@@ -1,5 +1,6 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
+import 'mapbox-gl/dist/mapbox-gl.css'
 import './index.css'
 import App from './App.tsx'
